api/requirements-api.txt
fastapi==0.110.0
uvicorn==0.27.1
boto3==1.34.0
pyathena==3.0.0
python-dotenv==1.0.1